+++ wdn/templates_5.3/js-src/datepickers.babel.js
@@ -1,7 +1,0 @@
-require(['dcf-datepicker'], (datePickerModule) => {
-  //Date Pickers
-  const datePickers = document.querySelectorAll('.dcf-datepicker');
-  datePickers.forEach(function (dp) {
-    new datePickerModule.DCFDatepicker(dp);
-  });
-});