+++ wdn/templates_6.0/js-src.old/card-as-link.babel.js
@@ -1,5 +1,0 @@
-require (['dcf-cardAsLink'], (cardAsLinkModule) => {
-  const cards = document.querySelectorAll('.dcf-card-as-link');
-  const cardAsLink = new cardAsLinkModule.DCFCardAsLink(cards);
-  cardAsLink.initialize();
-});